test: add deprecated tests & remove time module in tests
